M5.5 chunk 3.5: enum parameters in the FFI vocabulary.
#[buck_export] fns take #[buck_enum] types by value: the enum crosses as its repr scalar with generated discriminant validation (out-of-range from C# is InvalidArgument naming the enum and value, never a transmuted invalid enum), via the new BuckEnum trait every #[buck_enum] type implements (repr type + validated buck_from_raw/buck_raw). The generated C# is typed at both layers -- the raw import and the Native.* wrapper take the real C# enum.

First consumer: buck_log takes LogLevel (raw ABI unchanged; repr i32), its hand int-to-level match is gone, and Log.cs drops its cast. Chunk 4's public log-sink trait interface is born typed. One fence relocated: a bare struct param is now syntactically indistinguishable from an enum, so it fails on the BuckEnum trait bound instead of a prose macro error (trybuild case re-blessed with the sandbox-vs-real-world note).

Pinned by an expansion test (repr projection, validation glue, EnumRepr raw metadata), runtime probes for u32 and i32-with-negative-discriminant enums (valid echo + invalid-discriminant loudness), and the untouched preconfigure test proving the buck_log ABI held.

+++ src/core/cs/Log.cs
@@ -40,6 +40,6 @@ public static class Log
     private static void Emit(LogLevel level, string message)
     {
         // FfiCall rethrows a throwing log sink's exception right here -- at the causal call site, which is the point of exit-drain delivery.
-        Native.Log((int)level, message);
+        Native.Log(level, message);
     }
 }
